fix(remote): harden mobile prompt delivery
