+++ mobile/src/remote/timeline.ts
@@ -5,6 +5,7 @@
 export {
   appendUserMessage,
   applyStreamEvent,
+  commitAcknowledgedUserMessage,
   emptyTimeline,
   markApprovalDecision,
   mergeHistoryAttachments,
